--- crates/cel-fork/cel-derive/tests/expand/basic_struct.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct BasicStruct {
-    name: String,
-    age: i32,
-    active: bool,
-}
