--- crates/cel-fork/cel-derive/tests/expand/newtype_struct.rs
@@ -1,4 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct ExtAuthzDynamicMetadata(serde_json::Map<String, serde_json::Value>);
