--- crates/cel-fork/cel-derive/tests/expand/unit_enum.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub enum BackendType {
-    Http,
-    Grpc,
-    WebSocket,
-}
