--- crates/cel-fork/cel-derive/tests/expand/unit_enum_rename.rs
@@ -1,11 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub enum BackendProtocol {
-    #[dynamic(rename = "http")]
-    Http,
-    #[dynamic(rename = "grpc")]
-    Grpc,
-    #[dynamic(rename = "ws")]
-    WebSocket,
-}
