--- crates/cel-fork/cel-derive/tests/expand/with_flatten.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct WithFlatten {
-    key: String,
-    #[dynamic(flatten)]
-    metadata: serde_json::Value,
-}
