--- crates/cel-fork/cel-derive/tests/expand/with_option.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct WithOption {
-    required: String,
-    optional: Option<String>,
-    nested_optional: Option<i32>,
-}
