--- crates/cel-fork/cel-derive/tests/expand/with_rename.rs
@@ -1,9 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct WithRename {
-    #[dynamic(rename = "firstName")]
-    first_name: String,
-    #[dynamic(rename = "lastName")]
-    last_name: String,
-}
