--- crates/cel-fork/cel-derive/tests/expand/with_rename_all_camel.rs
@@ -1,9 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-#[dynamic(rename_all = "camelCase")]
-pub struct WithRenameAllCamel {
-    user_name: String,
-    user_age: i32,
-    is_active: bool,
-}
