--- crates/cel-fork/cel-derive/tests/expand/with_rename_all_lowercase.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-#[dynamic(rename_all = "lowercase")]
-pub struct WithRenameAllLower {
-    UserName: String,
-    UserAge: i32,
-}
