--- crates/cel-fork/cel-derive/tests/expand/with_serde.rs
@@ -1,10 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct WithSerde {
-    required: String,
-    #[serde(skip)]
-    internal_id: u64,
-    #[serde(rename = "custom_name")]
-    original_name: String,
-}
