--- crates/cel-fork/cel-derive/tests/expand/with_serde_rename_all.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-#[serde(rename_all = "camelCase")]
-pub struct WithSerdeRenameAll {
-    user_name: String,
-    user_id: i32,
-}
