--- crates/cel-fork/cel-derive/tests/expand/with_skip.rs
@@ -1,8 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct WithSkip {
-    public_field: String,
-    #[dynamic(skip)]
-    internal_field: u64,
-}
