--- crates/cel-fork/cel-derive/tests/expand/with_skip_serializing_if.rs
@@ -1,10 +1,0 @@
-use cel_derive::DynamicType;
-
-#[derive(DynamicType)]
-pub struct WithSkipSerializingIf<'a> {
-    required: &'a str,
-    #[dynamic(skip_serializing_if = "Option::is_none")]
-    optional: Option<&'a str>,
-    #[serde(skip_serializing_if = "str::is_empty")]
-    name: &'a str,
-}
